+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Task life cycle:
     * Support both pre-pinning and post-pinning of dependencies (\#2761).
     * Drop `FRACTAL_MAX_PIP_VERSION` configuration variable (\#2766).
+* Testing:
+    * Add `out_of_memory` mock task (\#2770).
 
 # 2.15.9